requirements.txt
greenlet==1.1.2
psycopg2-binary==2.9.2
python-dotenv==0.19.2
SQLAlchemy==1.4.27
